refactor(location): distinguish between street_name_part and street_name
