+++ src/locales/az/location/index.ts
@@ -12,7 +12,7 @@ import postcode from './postcode';
 import secondary_address from './secondary_address';
 import state from './state';
 import street_address from './street_address';
-import street_name from './street_name';
+import street_name_part from './street_name_part';
 import street_pattern from './street_pattern';
 import street_suffix from './street_suffix';
 
@@ -26,7 +26,7 @@ const location: LocationDefinition = {
   secondary_address,
   state,
   street_address,
-  street_name,
+  street_name_part,
   street_pattern,
   street_suffix,
 };
