+++ src/locales/az/location/street_pattern.ts
@@ -1,4 +1,1 @@
-export default [
-  '{{location.street_suffix}} {{location.street_name}}',
-  '{{location.street_name}} {{location.street_suffix}}',
-];
+export default ['{{location.street_name_part}} {{location.street_suffix}}'];
