+++ src/locales/es_MX/location/street_name_part.ts
