+++ src/locales/es_MX/location/street_prefix.ts
@@ -1,0 +1,25 @@
+export default [
+  'Calle',
+  'Avenida',
+  'Av.',
+  'Boulevard',
+  'Blvd.',
+  'Bulevar',
+  'Calzada',
+  'Calz.',
+  'Carretera',
+  'Carr.',
+  'Camino',
+  'Privada',
+  'Priv.',
+  'Cerrada',
+  'Cda.',
+  'Andador',
+  'Retorno',
+  'Prolongación',
+  'Prol.',
+  'Circuito',
+  'Paseo',
+  'Eje',
+  'Viaducto',
+];
