--- src/locales/es_MX/location/street_suffix.ts
@@ -1,72 +1,0 @@
-export default [
-  'Aldea',
-  'Apartamento',
-  'Arrabal',
-  'Arroyo',
-  'Avenida',
-  'Bajada',
-  'Barranco',
-  'Barrio',
-  'Bloque',
-  'Calle',
-  'Calleja',
-  'Camino',
-  'Carretera',
-  'Caserio',
-  'Colegio',
-  'Colonia',
-  'Conjunto',
-  'Cuesta',
-  'Chalet',
-  'Edificio',
-  'Entrada',
-  'Escalinata',
-  'Explanada',
-  'Extramuros',
-  'Extrarradio',
-  'Ferrocarril',
-  'Glorieta',
-  'Gran Subida',
-  'Grupo',
-  'Huerta',
-  'Jardines',
-  'Lado',
-  'Lugar',
-  'Manzana',
-  'Masía',
-  'Mercado',
-  'Monte',
-  'Muelle',
-  'Municipio',
-  'Parcela',
-  'Parque',
-  'Partida',
-  'Pasaje',
-  'Paseo',
-  'Plaza',
-  'Poblado',
-  'Polígono',
-  'Prolongación',
-  'Puente',
-  'Puerta',
-  'Quinta',
-  'Ramal',
-  'Rambla',
-  'Rampa',
-  'Riera',
-  'Rincón',
-  'Ronda',
-  'Rua',
-  'Salida',
-  'Sector',
-  'Sección',
-  'Senda',
-  'Solar',
-  'Subida',
-  'Terrenos',
-  'Torrente',
-  'Travesía',
-  'Urbanización',
-  'Vía',
-  'Vía Pública',
-];
