+++ src/locales/he/location/street_name_part.ts
