+++ src/locales/he/location/street_pattern.ts
@@ -1,1 +1,1 @@
-export default ['{{location.street_prefix}} {{location.street_name}}'];
+export default ['{{location.street_prefix}} {{location.street_name_part}}'];
