+++ src/locales/hu/location/street_name_part.ts
