+++ src/locales/hu/location/street_pattern.ts
@@ -1,5 +1,5 @@
 export default [
   '{{person.firstName}} {{location.street_suffix}}',
   '{{person.lastName}} {{location.street_suffix}}',
-  '{{location.street_name}} {{location.street_suffix}}',
+  '{{location.street_name_part}} {{location.street_suffix}}',
 ];
