+++ src/locales/ka_GE/location/street_pattern.ts
@@ -1,1 +1,1 @@
-export default ['{{location.street_name}} {{location.street_suffix}}'];
+export default ['{{location.street_name_part}} {{location.street_suffix}}'];
