+++ src/locales/lv/location/street_name_part.ts
