+++ src/locales/nb_NO/location/index.ts
@@ -16,7 +16,7 @@ import postcode from './postcode';
 import secondary_address from './secondary_address';
 import state from './state';
 import street_address from './street_address';
-import street_name from './street_name';
+import street_name_part from './street_name_part';
 import street_pattern from './street_pattern';
 import street_prefix from './street_prefix';
 import street_suffix from './street_suffix';
@@ -35,7 +35,7 @@ const location: LocationDefinition = {
   secondary_address,
   state,
   street_address,
-  street_name,
+  street_name_part,
   street_pattern,
   street_prefix,
   street_suffix,
