+++ src/locales/nb_NO/location/street_name_part.ts
