+++ src/locales/nb_NO/location/street_pattern.ts
@@ -1,6 +1,6 @@
 export default [
-  '{{location.street_name}}{{location.street_suffix}}',
-  '{{location.street_prefix}} {{location.street_name}}{{location.street_suffix}}',
+  '{{location.street_name_part}}{{location.street_suffix}}',
+  '{{location.street_prefix}} {{location.street_name_part}}{{location.street_suffix}}',
   '{{person.firstName}}{{location.common_street_suffix}}',
   '{{person.lastName}}{{location.common_street_suffix}}',
 ];
