+++ src/locales/ro/location/index.ts
@@ -12,7 +12,7 @@ import secondary_address from './secondary_address';
 import state from './state';
 import state_abbr from './state_abbr';
 import street_address from './street_address';
-import street_name from './street_name';
+import street_name_part from './street_name_part';
 import street_pattern from './street_pattern';
 import street_suffix from './street_suffix';
 
@@ -26,7 +26,7 @@ const location: LocationDefinition = {
   state,
   state_abbr,
   street_address,
-  street_name,
+  street_name_part,
   street_pattern,
   street_suffix,
 };
