+++ src/locales/ro/location/street_pattern.ts
@@ -1,1 +1,1 @@
-export default ['{{location.street_suffix}} {{location.street_name}}'];
+export default ['{{location.street_suffix}} {{location.street_name_part}}'];
