+++ src/locales/ru/location/street_name_part.ts
