+++ src/locales/sv/location/street_name_part.ts
