+++ src/locales/uk/location/street_name_part.ts
