+++ src/locales/uk/location/street_pattern.ts
@@ -1,4 +1,4 @@
 export default [
-  '{{location.street_prefix}} {{location.street_name}}',
-  '{{location.street_name}} {{location.street_suffix}}',
+  '{{location.street_prefix}} {{location.street_name_part}}',
+  '{{location.street_name_part}} {{location.street_suffix}}',
 ];
